+++ app/src/main/java/com/pzbproduction/ballerzfootball/viewmodel/FixturesFragmentViewModel.kt
@@ -17,11 +17,28 @@ class FixturesFragmentViewModel : ViewModel() {
     private var apiDataClass: MutableLiveData<ArrayList<ApiDataClass>> = MutableLiveData()
     private var logosOfHomeClubs: MutableLiveData<ArrayList<Bitmap>> = MutableLiveData()
     private var logosOfAwayClubs: MutableLiveData<ArrayList<Bitmap>> = MutableLiveData()
+    private var logosOfHomeClubsFromSpecificDate: MutableLiveData<ArrayList<Bitmap>> =
+        MutableLiveData()
+    private var logosOfAwayClubsFromSpecificDate: MutableLiveData<ArrayList<Bitmap>> =
+        MutableLiveData()
+    private var apiDataClassFromSpinner: MutableLiveData<ArrayList<ApiDataClass>> =
+        MutableLiveData()
+    private var matchesBySpecificDate: MutableLiveData<ArrayList<ApiDataClass>> =
+        MutableLiveData()
+    private var isLoading: MutableLiveData<Boolean> = MutableLiveData()
+    private var hasLoaded: MutableLiveData<Boolean> = MutableLiveData()
     private var repository = Repository()
 
     val getApiDataClass: LiveData<ArrayList<ApiDataClass>> get() = apiDataClass
     val getLogosOfHomeClubs: LiveData<ArrayList<Bitmap>> get() = logosOfHomeClubs
     val getLogosOfAwayClubs: LiveData<ArrayList<Bitmap>> get() = logosOfAwayClubs
+    val getLogosOfHomeClubsFromSpecificDate: LiveData<ArrayList<Bitmap>> get() = logosOfHomeClubsFromSpecificDate
+    val getLogosOfAwayClubsFromSpecificDate: LiveData<ArrayList<Bitmap>> get() = logosOfAwayClubsFromSpecificDate
+    val getApiDataClassFromSpinner: LiveData<ArrayList<ApiDataClass>> get() = apiDataClassFromSpinner
+    val getMatchesBySpecificDate: LiveData<ArrayList<ApiDataClass>> get() = matchesBySpecificDate
+
+    val getIsLoading: LiveData<Boolean> get() = isLoading
+    val getHasLoaded: LiveData<Boolean> get() = hasLoaded
 
     fun getTodaysMatch(map: HashMap<String, String>) {
         viewModelScope.launch(Dispatchers.IO) {
@@ -40,6 +57,47 @@ class FixturesFragmentViewModel : ViewModel() {
         viewModelScope.launch(Dispatchers.IO) {
             logosOfAwayClubs.postValue(repository.getLogosOfAwayClubs())
         }
+    }
+
+    fun getLogosOfHomeClubFromSpecificDate(){
+        viewModelScope.launch(Dispatchers.IO) {
+            logosOfHomeClubsFromSpecificDate.postValue(repository.getLogosOfHomeClubsFromSpecificDate())
+        }
+    }
+
+    fun getLogosOfAwayClubFromSpecificDate(){
+        viewModelScope.launch(Dispatchers.IO) {
+            logosOfAwayClubsFromSpecificDate.postValue(repository.getLogosOfAwayClubsFromSpecificDate())
+        }
+    }
+
+
+    fun getApiDataClassFromSpinner(map: HashMap<String, String>) {
+        viewModelScope.launch(Dispatchers.IO) {
+            apiDataClassFromSpinner.postValue(repository.getMatchesBySelectionFromSpinner(map))
+        }
+    }
+
+    fun getMatchesBySpecificDate(startDate: String, endDate: String, map: HashMap<String, String>) {
+        viewModelScope.launch(Dispatchers.IO) {
+            matchesBySpecificDate.postValue(
+                repository.getMatchBySpecificDate(
+                    startDate,
+                    endDate,
+                    map
+                )
+            )
+        }
+    }
+
+
+    fun getIsLoading() {
+        isLoading.postValue(repository.getIsLoading())
+
+    }
+
+    fun getHasLoaded() {
+        hasLoaded.postValue(repository.getHasLoaded())
     }