+++ ChangeLog.md
@@ -2,6 +2,11 @@ CHANGELOG
 -----------------------
 -----------------------
 
+ver 1.1.1 (12th Apr 2023)
+-------------------------------
+--> Search based on league selected in spinner
+--> Added other dates other than today
+
 ver 1.0.1 (11th Apr 2023)
 --------------------------------
 --> ChangeLog added